# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Roamwise Travel Experience Optimization API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trip found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/analyze:
    post:
      operationId: analyzeTrip
      tags: [trips]
      summary: AI analysis of trip destinations and strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripAnalysis"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/generate:
    post:
      operationId: generateItinerary
      tags: [trips]
      summary: Generate full AI itinerary for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Itinerary generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /trips/{id}/modify:
    post:
      operationId: modifyItinerary
      tags: [trips]
      summary: Modify existing itinerary via natural language request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripModificationInput"
      responses:
        "200":
          description: Itinerary modified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripModification"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TravelerProfile:
      type: object
      properties:
        interests:
          type: array
          items:
            type: string
        travel_style:
          type: string
        preferences:
          type: array
          items:
            type: string
      required:
        - interests
        - travel_style
        - preferences
    TripInput:
      type: object
      properties:
        destination:
          type: string
        starting_location:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        traveler_count:
          type: integer
        budget:
          type: number
        budget_preference:
          type: string
        traveler_profile:
          $ref: "#/components/schemas/TravelerProfile"
      required:
        - destination
        - starting_location
        - start_date
        - end_date
        - traveler_count
        - budget
        - budget_preference
        - traveler_profile
    DestinationScore:
      type: object
      properties:
        name:
          type: string
        score:
          type: integer
        reasoning:
          type: string
        drawbacks:
          type: string
        recommended_nights:
          type: integer
          minimum: 1
        why_selected:
          type: string
        experience_highlights:
          type: array
          items:
            type: string
        estimated_cost_usd:
          type: number
          minimum: 0
      required:
        - name
        - score
        - reasoning
        - drawbacks
...
    RouteStop:
      type: object
      properties:
        location:
          type: string
        nights:
          type: integer
        transport_to_next:
          type: ["string", "null"]
        duration_hours:
          type: ["number", "null"]
      required:
        - location
        - nights
        - transport_to_next
        - duration_hours
    DailyActivity:
      type: object
      properties:
        morning:
          type: string
        afternoon:
          type: string
        evening:
          type: string
        food_recommendation:
          type: string
        transport:
          type: string
        estimated_cost:
          type: number
      required:
        - morning
        - afternoon
        - evening
        - food_recommendation
        - transport
        - estimated_cost
    DailySchedule:
      type: object
      properties:
        day:
          type: integer
        date:
          type: string
        location:
          type: string
        activities:
          $ref: "#/components/schemas/DailyActivity"
      required:
        - day
        - date
        - location
        - activities
    DailyActivityPart:
      type: object
      properties:
        activity:
          type: string
        description:
          type: string
        estimated_cost_usd:
          type: number
          minimum: 0
      required:
        - activity
        - description
        - estimated_cost_usd
    DailyTransportation:
      type: object
      properties:
        mode:
          type: string
        details:
          type: string
        duration:
          type: string
      required:
        - mode
        - details
        - duration
    DailyItinerary:
      type: object
      properties:
        day:
          type: integer
        date:
          type: string
        location:
          type: string
        morning:
          $ref: "#/components/schemas/DailyActivityPart"
        afternoon:
          $ref: "#/components/schemas/DailyActivityPart"
        evening:
          $ref: "#/components/schemas/DailyActivityPart"
        food_recommendations:
          type: array
          items:
            type: string
        transportation:
          $ref: "#/components/schemas/DailyTransportation"
        estimated_daily_cost_usd:
          type: number
          minimum: 0
        weather:
          $ref: "#/components/schemas/WeatherSummary"
      required:
        - day
        - date
        - location
        - morning
        - afternoon
        - evening
        - food_recommendations
        - transportation
        - estimated_daily_cost_usd
    BudgetItem:
      type: object
      properties:
        category:
          type: string
        estimated_amount:
          type: number
        description:
          type: string
        source_metadata:
          $ref: "#/components/schemas/DataSourceMetadata"
      required:
        - category
        - estimated_amount
        - description
    BudgetSummary:
      type: object
      properties:
        currency:
          type: string
          enum: ["USD"]
        total_budget:
          type: number
          minimum: 0
        flights_estimated:
          type: number
          minimum: 0
        flights_live:
          type: number
          minimum: 0
        accommodation_estimated:
          type: number
          minimum: 0
        transportation_estimated:
          type: number
          minimum: 0
        food_estimated:
          type: number
          minimum: 0
        activities_estimated:
          type: number
          minimum: 0
        total_estimated:
          type: number
        remaining_budget:
          type: number
        estimates_only:
          type: boolean
      required:
        - currency
        - total_budget
        - flights_estimated
        - accommodation_estimated
        - transportation_estimated
        - food_estimated
        - activities_estimated
        - total_estimated
        - remaining_budget
        - estimates_only
    DataSourceMetadata:
      type: object
      properties:
        provider:
          type: string
        data_type:
          type: string
        retrieved_at:
          oneOf:
            - type: string
            - type: "null"
        freshness:
          type: string
        is_live:
          type: boolean
        label:
          type: string
          enum: ["LIVE", "ESTIMATED", "ROAMWISE RECOMMENDED", "FALLBACK"]
      required:
        - provider
        - data_type
        - retrieved_at
        - freshness
        - is_live
        - label
...
    WeatherSummary:
      type: object
      properties:
        location:
          type: string
        date:
          type: string
        min_temperature_c:
          type: number
        max_temperature_c:
          type: number
        precipitation_probability:
          oneOf:
            - type: number
            - type: "null"
        weather_code:
          type: integer
        description:
          type: string
        source_metadata:
          $ref: "#/components/schemas/DataSourceMetadata"
      required:
        - location
        - date
        - min_temperature_c
        - max_temperature_c
        - precipitation_probability
        - weather_code
        - description
        - source_metadata
    LiveDataBundle:
      type: object
      properties:
        flight_search:
          $ref: "#/components/schemas/FlightSearch"
        weather:
          type: array
          items:
            $ref: "#/components/schemas/WeatherSummary"
        refreshed_at:
          type: string
        planning_note:
          type: string
      required:
        - flight_search
        - weather
        - refreshed_at
    TradeoffItem:
      type: object
      properties:
        description:
          type: string
        impact:
          type: string
      required:
        - description
        - impact
    TripRecommendation:
      type: object
      properties:
        recommendation:
          type: string
        reason:
          type: string
        expected_benefit:
          type: string
        tradeoff:
          type: string
      required:
        - recommendation
        - reason
        - expected_benefit
        - tradeoff
    AlternativeStrategy:
      type: object
      properties:
        name:
          type: string
        best_for:
          type: array
          items:
            type: string
        changes:
          type: string
        gains:
          type: string
        sacrifices:
          type: string
        estimated_budget_impact_usd:
          type: number
        pacing_impact:
          type: string
        experience_match_impact:
          type: string
        recommended:
          type: boolean
      required:
        - name
        - best_for
        - changes
        - gains
        - sacrifices
        - estimated_budget_impact_usd
        - pacing_impact
        - experience_match_impact
        - recommended
    TripAnalysis:
      type: object
      properties:
        trip_strategy:
          type: string
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/DestinationScore"
        reasoning:
          type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/TripRecommendation"
        strategies:
          type: array
          items:
            $ref: "#/components/schemas/AlternativeStrategy"
      required:
        - trip_strategy
        - destinations
        - reasoning
    TripHealthSubScore:
      type: object
      description: One dimension of the Trip Health Score
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
        label:
          type: string
          enum: ["Excellent", "Good", "Fair", "Needs attention"]
        explanation:
          type: string
      required:
        - score
        - label
        - explanation
    TripHealthScore:
      type: object
      description: Structured multi-dimensional evaluation of the itinerary
      properties:
        overall:
          type: integer
          minimum: 0
          maximum: 100
        overall_label:
          type: string
          enum: ["Excellent", "Good", "Fair", "Needs attention"]
        experience_fit:
          $ref: "#/components/schemas/TripHealthSubScore"
        transportation_efficiency:
          $ref: "#/components/schemas/TripHealthSubScore"
        budget_efficiency:
          $ref: "#/components/schemas/TripHealthSubScore"
        uniqueness:
          $ref: "#/components/schemas/TripHealthSubScore"
        pacing:
          $ref: "#/components/schemas/TripHealthSubScore"
      required:
        - overall
        - overall_label
        - experience_fit
        - transportation_efficiency
        - budget_efficiency
        - uniqueness
        - pacing
    Itinerary:
      type: object
      properties:
        id:
          type: string
        trip_id:
          type: string
        currency:
          type: string
          enum: ["USD"]
        total_days:
          type: integer
          minimum: 1
        total_nights:
          type: integer
          minimum: 1
        trip_strategy:
          type: string
        route:
          type: array
          items:
            $ref: "#/components/schemas/RouteStop"
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/DestinationScore"
        daily_itinerary:
          type: array
          items:
            $ref: "#/components/schemas/DailyItinerary"
        daily_schedule:
          type: array
          items:
            $ref: "#/components/schemas/DailySchedule"
        budget_breakdown:
          type: array
          items:
            $ref: "#/components/schemas/BudgetItem"
        budget_summary:
          $ref: "#/components/schemas/BudgetSummary"
        live_data:
          $ref: "#/components/schemas/LiveDataBundle"
        reasoning:
          type: string
        tradeoffs:
          type: array
          items:
            $ref: "#/components/schemas/TradeoffItem"
        health_score:
          oneOf:
            - $ref: "#/components/schemas/TripHealthScore"
            - type: "null"
        version:
          type: integer
        created_at:
          type: string
      required:
        - id
        - trip_id
        - currency
        - total_days
        - total_nights
        - trip_strategy
        - route
        - destinations
        - daily_itinerary
        - daily_schedule
        - budget_breakdown
        - budget_summary
        - reasoning
        - tradeoffs
        - health_score
        - version
        - created_at
    Trip:
      type: object
      properties:
        id:
          type: string
        destination:
          type: string
        starting_location:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        traveler_count:
          type: integer
        budget:
          type: number
        currency:
          type: string
          enum: ["USD"]
        budget_preference:
          type: string
        traveler_profile:
          $ref: "#/components/schemas/TravelerProfile"
        latest_itinerary:
          oneOf:
            - $ref: "#/components/schemas/Itinerary"
            - type: "null"
        created_at:
          type: string
      required:
        - id
        - destination
        - starting_location
        - start_date
        - end_date
        - traveler_count
        - budget
         - currency
        - budget_preference
        - traveler_profile
        - latest_itinerary
        - created_at
    ChangeMade:
      type: object
      properties:
        description:
          type: string
        type:
          type: string
      required:
        - description
        - type
    TripModificationInput:
      type: object
      properties:
        user_request:
          type: string
      required:
        - user_request
    TripModification:
      type: object
      properties:
        id:
          type: string
        trip_id:
          type: string
        user_request:
          type: string
        changes_made:
          type: array
          items:
            $ref: "#/components/schemas/ChangeMade"
        reasoning:
          type: string
        score_before:
          oneOf:
            - $ref: "#/components/schemas/TripHealthScore"
            - type: "null"
        score_after:
          oneOf:
            - $ref: "#/components/schemas/TripHealthScore"
            - type: "null"
        itinerary:
          $ref: "#/components/schemas/Itinerary"
        created_at:
          type: string
      required:
        - id
        - trip_id
        - user_request
        - changes_made
        - reasoning
        - score_before
        - score_after
        - itinerary
        - created_at